Move the validate_seal method to the VM class.
